Add binutils to base system image
binutils is a base gnu userland toolset normally its included with coreutils as a dependency, but I guess with void its not. It has the both `ld` linkers, as well as `ar`, `strings`, `objdump`, etc. It's only 3mb, and of such a base level that we should probably include it.  When all else fails